Moved government page to its own module
- Preparing for future pages to be added to the module
- Also generally cleaner this way
