+++ RuddockWebsite/modules/government/__init__.py
@@ -1,0 +1,4 @@
+import flask
+blueprint = flask.Blueprint('government', __name__, template_folder='templates')
+
+import RuddockWebsite.modules.government.routes
