+++ RuddockWebsite/tests/government/test_government_routes.py
@@ -1,0 +1,14 @@
+"""
+Tests routes in the government module.
+"""
+
+import flask
+import httplib
+
+from RuddockWebsite.tests.fixtures import client
+
+def test_government(client):
+  """ Tests the /government route. """
+  response = client.get(flask.url_for('government.government_home'))
+  assert response.status_code == httplib.OK
+
